// abci/src/proto-ts/cometbft/abci/v1/types.ts
export interface EchoRequest {
  message: string;
}

/** FlushRequest is a request to flush the write buffer. */
export interface FlushRequest {
}

/** InfoRequest is a request for the ABCI application version. */
export interface InfoRequest {
  version: string;
  block_version: number;
  p2p_version: number;
  abci_version: string;
}
// ...
export interface CheckTxRequest {
  tx: Uint8Array;
  type: CheckTxType;
}

/** CommitRequest is a request to commit the pending application state. */
export interface CommitRequest {
}
// ...
export interface ApplySnapshotChunkRequest {
  index: number;
  chunk: Uint8Array;
  sender: string;
}
// ...
export interface EchoResponse {
  message: string;
}

/** FlushResponse indicates that the write buffer was flushed. */
export interface FlushResponse {
}

/** InfoResponse contains the ABCI application version information. */
export interface InfoResponse {
  data: string;
  version: string;
  app_version: number;
  last_block_height: number;
  last_block_app_hash: Uint8Array;
}

/**
 * InitChainResponse contains the ABCI application's hash and updates to the
 * validator set and/or the consensus params, if any.
 */
export interface InitChainResponse {
  consensus_params?: ConsensusParams | undefined;
  validators: ValidatorUpdate[];
  app_hash: Uint8Array;
}
// ...
export interface CheckTxResponse {
  code: number;
  data: Uint8Array;
  /** nondeterministic */
  log: string;
  /** nondeterministic */
  info: string;
  gas_wanted: number;
  gas_used: number;
  /** nondeterministic */
  events: Event[];
  codespace: string;
}

/** CommitResponse indicates how much blocks should CometBFT retain. */
export interface CommitResponse {
  retain_height: number;
}
// ...
export interface ApplySnapshotChunkResponse {
  result: ApplySnapshotChunkResult;
  /** Chunks to refetch and reapply */
  refetch_chunks: number[];
  /** Chunk senders to reject and ban */
  reject_senders: string[];
}
// ...
export interface ExtendVoteResponse {
  vote_extension: Uint8Array;
}
// ...
export const CheckTxRequest: MessageFns<CheckTxRequest> = {
  encode(message: CheckTxRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.tx.length !== 0) {
      writer.uint32(10).bytes(message.tx);
    }
    if (message.type !== 0) {
      writer.uint32(24).int32(message.type);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CheckTxRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCheckTxRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tx = reader.bytes();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CheckTxRequest {
    return {
      tx: isSet(object.tx) ? bytesFromBase64(object.tx) : new Uint8Array(0),
      type: isSet(object.type) ? checkTxTypeFromJSON(object.type) : 0,
    };
  },

  toJSON(message: CheckTxRequest): unknown {
    const obj: any = {};
    if (message.tx.length !== 0) {
      obj.tx = base64FromBytes(message.tx);
    }
    if (message.type !== 0) {
      obj.type = checkTxTypeToJSON(message.type);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CheckTxRequest>, I>>(base?: I): CheckTxRequest {
    return CheckTxRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CheckTxRequest>, I>>(object: I): CheckTxRequest {
    const message = createBaseCheckTxRequest();
    message.tx = object.tx ?? new Uint8Array(0);
    message.type = object.type ?? 0;
    return message;
  },
};

function createBaseCommitRequest(): CommitRequest {
  return {};
}

export const CommitRequest: MessageFns<CommitRequest> = {
  encode(_: CommitRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CommitRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCommitRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): CommitRequest {
    return {};
  },

  toJSON(_: CommitRequest): unknown {
    const obj: any = {};
    return obj;
  },

  create<I extends Exact<DeepPartial<CommitRequest>, I>>(base?: I): CommitRequest {
    return CommitRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CommitRequest>, I>>(_: I): CommitRequest {
    const message = createBaseCommitRequest();
    return message;
  },
};
// ...
export const ApplySnapshotChunkRequest: MessageFns<ApplySnapshotChunkRequest> = {
  encode(message: ApplySnapshotChunkRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.index !== 0) {
      writer.uint32(8).uint32(message.index);
    }
    if (message.chunk.length !== 0) {
      writer.uint32(18).bytes(message.chunk);
    }
    if (message.sender !== "") {
      writer.uint32(26).string(message.sender);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplySnapshotChunkRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplySnapshotChunkRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.index = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.chunk = reader.bytes();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.sender = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplySnapshotChunkRequest {
    return {
      index: isSet(object.index) ? globalThis.Number(object.index) : 0,
      chunk: isSet(object.chunk) ? bytesFromBase64(object.chunk) : new Uint8Array(0),
      sender: isSet(object.sender) ? globalThis.String(object.sender) : "",
    };
  },

  toJSON(message: ApplySnapshotChunkRequest): unknown {
    const obj: any = {};
    if (message.index !== 0) {
      obj.index = Math.round(message.index);
    }
    if (message.chunk.length !== 0) {
      obj.chunk = base64FromBytes(message.chunk);
    }
    if (message.sender !== "") {
      obj.sender = message.sender;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplySnapshotChunkRequest>, I>>(base?: I): ApplySnapshotChunkRequest {
    return ApplySnapshotChunkRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplySnapshotChunkRequest>, I>>(object: I): ApplySnapshotChunkRequest {
    const message = createBaseApplySnapshotChunkRequest();
    message.index = object.index ?? 0;
    message.chunk = object.chunk ?? new Uint8Array(0);
    message.sender = object.sender ?? "";
    return message;
  },
};
// ...
export const CommitResponse: MessageFns<CommitResponse> = {
  encode(message: CommitResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.retain_height !== 0) {
      writer.uint32(24).int64(message.retain_height);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CommitResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCommitResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.retain_height = longToNumber(reader.int64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CommitResponse {
    return { retain_height: isSet(object.retain_height) ? globalThis.Number(object.retain_height) : 0 };
  },

  toJSON(message: CommitResponse): unknown {
    const obj: any = {};
    if (message.retain_height !== 0) {
      obj.retain_height = Math.round(message.retain_height);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CommitResponse>, I>>(base?: I): CommitResponse {
    return CommitResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CommitResponse>, I>>(object: I): CommitResponse {
    const message = createBaseCommitResponse();
    message.retain_height = object.retain_height ?? 0;
    return message;
  },
};
// ...
export const ApplySnapshotChunkResponse: MessageFns<ApplySnapshotChunkResponse> = {
  encode(message: ApplySnapshotChunkResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.result !== 0) {
      writer.uint32(8).int32(message.result);
    }
    writer.uint32(18).fork();
    for (const v of message.refetch_chunks) {
      writer.uint32(v);
    }
    writer.join();
    for (const v of message.reject_senders) {
      writer.uint32(26).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplySnapshotChunkResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplySnapshotChunkResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.result = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag === 16) {
            message.refetch_chunks.push(reader.uint32());

            continue;
          }

          if (tag === 18) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.refetch_chunks.push(reader.uint32());
            }

            continue;
          }

          break;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.reject_senders.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplySnapshotChunkResponse {
    return {
      result: isSet(object.result) ? applySnapshotChunkResultFromJSON(object.result) : 0,
      refetch_chunks: globalThis.Array.isArray(object?.refetch_chunks)
        ? object.refetch_chunks.map((e: any) => globalThis.Number(e))
        : [],
      reject_senders: globalThis.Array.isArray(object?.reject_senders)
        ? object.reject_senders.map((e: any) => globalThis.String(e))
        : [],
    };
  },

  toJSON(message: ApplySnapshotChunkResponse): unknown {
    const obj: any = {};
    if (message.result !== 0) {
      obj.result = applySnapshotChunkResultToJSON(message.result);
    }
    if (message.refetch_chunks?.length) {
      obj.refetch_chunks = message.refetch_chunks.map((e) => Math.round(e));
    }
    if (message.reject_senders?.length) {
      obj.reject_senders = message.reject_senders;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplySnapshotChunkResponse>, I>>(base?: I): ApplySnapshotChunkResponse {
    return ApplySnapshotChunkResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplySnapshotChunkResponse>, I>>(object: I): ApplySnapshotChunkResponse {
    const message = createBaseApplySnapshotChunkResponse();
    message.result = object.result ?? 0;
    message.refetch_chunks = object.refetch_chunks?.map((e) => e) || [];
    message.reject_senders = object.reject_senders?.map((e) => e) || [];
    return message;
  },
};